@@ -14,197 +15,119 @@
   },
   "scripts": {
     "prepare": "lefthook install",
-    "storybook": "storybook dev -p 6006",
-    "serve:storybook": "npx http-server ./storybook-static",
-    "build:storybook": "storybook build",
-    "prebuild": "pnpm --prefix packages/cli build && pnpm gen:tokens",
-    "build": "turbo run build --filter=./packages/**",
-    "build:depend": "tsx scripts/dependency-build",
-    "build:fast": "turbo run build:fast --filter=./packages/**",
-    "dev": "turbo run dev --filter=./packages/**",
-    "lint": "eslint . --max-warnings=0 --cache --flag unstable_ts_config",
-    "lint:inspect": "pnpm dlx @eslint/config-inspector",
-    "format": "prettier --write .",
-    "format:check": "prettier --check packages/**/** stories/**/** scripts/**/** --cache",
-    "format:write": "prettier --write packages/**/** stories/**/** scripts/**/** --cache",
-    "typecheck": "tsc --noEmit",
-    "typetrace": "tsc --noEmit --generateTrace ./trace --incremental false",
-    "typeanalyze": "npx analyze-trace ./trace --expandTypes=false",
-    "spellcheck": "cspell \"**/*.{js,jsx,mjs,ts,tsx,mts,md,mdx}\" --quiet",
-    "test": "vitest",
-    "test:watch": "vitest --watch",
-    "test:issue": "tsx scripts/coverage-issues && tsx scripts/a11y-issues",
-    "test:ci": "vitest run --coverage",
-    "test:ui": "vitest --coverage --watch --ui",
-    "test:a11y": "test-storybook",
-    "quality": "pnpm format:check && pnpm lint && pnpm typecheck && pnpm test",
-    "gen:css": "tsx scripts/generate-css && eslint packages/core/src/styles.ts --fix --flag unstable_ts_config",
-    "gen:tokens": "yamada-cli tokens packages/theme/src/index.ts --out packages/core/src/generated-theme.types.ts && eslint packages/core/src/generated-theme.types.ts --fix --flag unstable_ts_config",
-    "gen:component": "cross-env NODE_OPTIONS='--import tsx' plop component && pnpm install",
-    "gen:hook": "cross-env NODE_OPTIONS='--import tsx' plop hook && pnpm install",
-    "gen:docs": "pnpm --filter=./packages/components/** gen:docs",
-    "changelog": "tsx scripts/changelog",
-    "changelog:ci": "tsx scripts/changelog --current",
-    "changelog:latest": "tsx scripts/changelog --latest",
-    "version:dev": "changeset version --snapshot dev",
-    "version:next": "changeset version --snapshot next",
-    "release": "changeset publish",
-    "release:dev": "changeset publish --tag dev",
-    "release:next": "changeset publish --tag next",
-    "report:insights": "tsx scripts/report-insights",
-    "notification:release": "tsx scripts/notification-release",
-    "clean": "pnpm -r --parallel exec rimraf dist .turbo *.log node_modules",
+    "clean": "pnpm -r --parallel exec rimraf dist .turbo *.log .eslintcache coverage node_modules",
     "clean:all": "pnpm clean && pnpm rimraf .turbo *.log .eslintcache coverage storybook-static trace node_modules",
-    "docs:dev": "pnpm --prefix docs dev",
-    "docs:start": "pnpm --prefix docs start",
-    "docs:build": "pnpm --prefix docs build",
-    "docs:lint": "pnpm --prefix docs lint",
-    "docs:format": "pnpm --prefix docs format",
-    "docs:format:check": "pnpm --prefix docs format:check",
-    "docs:typecheck": "pnpm --prefix docs typecheck",
-    "docs:test": "pnpm --prefix docs test",
-    "docs:content": "pnpm --prefix docs gen:content",
-    "docs:changelog": "pnpm --prefix docs gen:changelog",
-    "docs:theme": "pnpm --prefix docs gen:theme",
-    "docs:styled-system": "pnpm --prefix docs gen:styled-system",
-    "docs:props": "pnpm --prefix docs gen:props",
-    "docs:search": "pnpm --prefix docs gen:search",
-    "docs:icon-tags": "pnpm --prefix docs gen:icon-tags",
-    "docs:contributors": "pnpm --prefix docs gen:contributors",
-    "docs:translate": "pnpm --prefix docs gen:translate",
-    "docs:tokens": "pnpm --prefix docs gen:tokens",
-    "docs:label": "pnpm --prefix docs gen:label",
-    "react:start": "pnpm --prefix examples/create-react-app start",
-    "react:build": "pnpm --prefix examples/create-react-app build",
-    "next-pages:dev": "pnpm --prefix examples/next/pages dev",
-    "next-pages:build": "pnpm --prefix examples/next/pages build",
-    "next-pages:start": "pnpm --prefix examples/next/pages start",
-    "next-app:dev": "pnpm --prefix examples/next/app dev",
-    "next-app:build": "pnpm --prefix examples/next/app build",
-    "next-app:start": "pnpm --prefix examples/next/app start",
-    "vite:dev": "pnpm --prefix examples/vite dev",
-    "vite:build": "pnpm --prefix examples/vite build",
-    "vite:preview": "pnpm --prefix examples/vite preview",
-    "remix:dev": "pnpm --prefix examples/remix dev",
-    "remix:build": "pnpm --prefix examples/remix build",
-    "remix:start": "pnpm --prefix examples/remix start",
-    "hono:dev": "pnpm --prefix examples/hono dev",
-    "hono:build": "pnpm --prefix examples/hono build"
+    "cli": "pnpm --prefix packages/cli",
+    "react": "pnpm --prefix packages/react",
+    "utils": "pnpm --prefix packages/utils",
+    "www": "pnpm --prefix www",
+    "next:pages": "pnpm --prefix playgrounds/next/pages",
+    "next:app": "pnpm --prefix playgrounds/next/app",
+    "vite": "pnpm --prefix playgrounds/vite",
+    "remix": "pnpm --prefix playgrounds/remix",
+    "hono": "pnpm --prefix playgrounds/hono",
+    "prebuild": "pnpm cli build && pnpm gen:tokens",
+    "build": "turbo run build --filter=./packages/**",
+    "lint": "turbo run lint --filter=./packages/**",
+    "lint:inspect": "pnpm dlx @eslint/config-inspector",
+    "format": "turbo run format --filter=./packages/**",
+    "typecheck": "turbo run typecheck --filter=./packages/**",
+    "test": "turbo run test --filter=./packages/**",
+    "spellcheck": "cspell \"**/*.{js,jsx,mjs,ts,tsx,mts,md,mdx}\" --quiet",
+    "quality": "pnpm format && pnpm lint && pnpm typecheck && pnpm test",
+    "gen:styles": "tsx scripts/generate/styles",
+    "gen:tokens": "yamada-cli tokens packages/react/src/theme/index.ts --internal --out packages/react/src/core/generated-theme-tokens.types.ts && eslint packages/react/src/core/generated-theme-tokens.types.ts --fix",
+    "gen:icons": "tsx scripts/generate/icons",
+    "gen:props": "tsx scripts/generate/props",
+    "changelog": "tsx scripts/changelog",
+    "release": "changeset publish",
+    "release:dev": "changeset version --snapshot dev && changeset publish --tag dev",
+    "release:next": "changeset version --snapshot next && changeset publish --tag next"
   },
   "dependencies": {
+    "@babel/parser": "^7.27.5",
+    "@babel/traverse": "^7.27.4",
     "@changesets/changelog-github": "^0.5.1",
-    "@changesets/cli": "^2.28.1",
-    "@changesets/get-release-plan": "^4.0.8",
+    "@changesets/cli": "^2.29.5",
+    "@changesets/get-release-plan": "^4.0.13",
     "@changesets/types": "^6.1.0",
-    "@clack/prompts": "^0.10.0",
-    "@commitlint/cli": "^19.7.1",
-    "@commitlint/config-conventional": "^19.7.1",
-    "@cspell/eslint-plugin": "^8.17.4",
-    "@emotion/react": "11.11.0",
-    "@emotion/styled": "11.11.0",
-    "@eslint/compat": "^1.2.6",
-    "@faker-js/faker": "^9.5.0",
-    "@fortawesome/fontawesome-svg-core": "^6.7.2",
-    "@fortawesome/free-brands-svg-icons": "^6.7.2",
-    "@fortawesome/free-regular-svg-icons": "^6.7.2",
-    "@fortawesome/free-solid-svg-icons": "^6.7.2",
-    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@commitlint/cli": "^19.8.1",
+    "@commitlint/config-conventional": "^19.8.1",
+    "@cspell/eslint-plugin": "^9.1.2",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@eslint/compat": "^1.3.1",
     "@koralle/vitest-axe": "^0.1.1",
-    "@lucide/lab": "^0.1.2",
-    "@octokit/request-error": "^6.1.7",
-    "@octokit/rest": "^21.1.1",
-    "@storybook/addon-a11y": "^8.5.8",
-    "@storybook/addon-backgrounds": "^8.5.8",
-    "@storybook/addon-docs": "^8.5.8",
-    "@storybook/addon-measure": "^8.5.8",
-    "@storybook/addon-storysource": "^8.5.8",
-    "@storybook/addon-viewport": "^8.5.8",
-    "@storybook/blocks": "^8.5.8",
-    "@storybook/cli": "^8.5.8",
-    "@storybook/preview-api": "^8.5.8",
-    "@storybook/react": "^8.5.8",
-    "@storybook/react-vite": "^8.5.8",
-    "@storybook/test-runner": "^0.22.0",
-    "@storybook/theming": "^8.5.8",
+    "@mdn/browser-compat-data": "^6.0.26",
+    "@octokit/request-error": "^7.0.0",
+    "@octokit/rest": "^22.0.0",
+    "@storybook/addon-a11y": "^9.0.13",
+    "@storybook/addon-docs": "^9.0.13",
+    "@storybook/cli": "^9.0.13",
+    "@storybook/react-vite": "^9.0.13",
     "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.2.0",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/babel__traverse": "^7.20.7",
     "@types/jsdom": "^21.1.7",
-    "@types/node": "^22.13.4",
-    "@types/react": "^18.3.18",
-    "@types/react-dom": "^18.3.5",
-    "@typescript-eslint/parser": "^8.24.1",
+    "@types/node": "^24.0.4",
+    "@typescript-eslint/parser": "^8.35.0",
     "@typescript/analyze-trace": "^0.10.1",
-    "@vitejs/plugin-react-swc": "^3.8.0",
-    "@vitest/coverage-v8": "^3.0.6",
-    "@vitest/eslint-plugin": "^1.1.31",
-    "@vitest/ui": "^3.0.6",
-    "@yamada-ui/calendar": "workspace:*",
-    "@yamada-ui/carousel": "workspace:*",
-    "@yamada-ui/charts": "workspace:*",
+    "@vitejs/plugin-react-swc": "^3.10.2",
+    "@vitest/coverage-v8": "^3.2.4",
+    "@vitest/eslint-plugin": "^1.2.7",
+    "@vitest/ui": "^3.2.4",
     "@yamada-ui/cli": "workspace:*",
-    "@yamada-ui/core": "workspace:*",
-    "@yamada-ui/dropzone": "workspace:*",
-    "@yamada-ui/fontawesome": "workspace:*",
-    "@yamada-ui/lucide": "workspace:*",
-    "@yamada-ui/markdown": "workspace:*",
-    "@yamada-ui/next": "workspace:*",
     "@yamada-ui/react": "workspace:*",
-    "@yamada-ui/table": "workspace:*",
-    "@yamada-ui/test": "workspace:*",
-    "@yamada-ui/theme": "workspace:*",
     "@yamada-ui/utils": "workspace:*",
     "axe-playwright": "^2.1.0",
-    "axios": "^1.7.9",
-    "chalk": "^5.4.1",
-    "commander": "^13.1.0",
-    "cross-env": "^7.0.3",
+    "commander": "^14.0.0",
     "csstype": "^3.1.3",
     "dayjs": "^1.11.13",
-    "dotenv": "^16.4.7",
-    "eslint": "^9.20.1",
-    "eslint-config-prettier": "^10.0.1",
+    "dotenv": "^16.5.0",
+    "eslint": "^9.29.0",
+    "eslint-config-prettier": "^10.1.5",
     "eslint-import-resolver-node": "^0.3.9",
-    "eslint-import-resolver-typescript": "^3.8.3",
-    "eslint-plugin-import": "^2.31.0",
+    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-import-replace": "^1.0.0",
     "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-perfectionist": "^3.9.1",
-    "eslint-plugin-react": "^7.37.4",
-    "eslint-plugin-react-hooks": "^5.1.0",
-    "eslint-plugin-testing-library": "^7.1.1",
+    "eslint-plugin-perfectionist": "^4.15.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-storybook": "9.0.13",
+    "eslint-plugin-testing-library": "^7.5.3",
     "eslint-plugin-unused-imports": "^4.1.4",
     "execa": "9.3.1",
     "find-packages": "^10.0.4",
-    "glob": "^11.0.1",
-    "globals": "^16.0.0",
+    "glob": "^11.0.3",
+    "globals": "^16.2.0",
     "jiti": "^2.4.2",
-    "jsdom": "^26.0.0",
-    "lefthook": "^1.10.10",
-    "list-it": "^1.3.13",
-    "lucide-react": "^0.475.0",
+    "jsdom": "^26.1.0",
+    "lefthook": "^1.11.14",
+    "lucide-react": "^0.515.0",
     "node-fetch": "^3.3.2",
-    "plop": "^4.0.1",
-    "prettier": "^3.5.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-hook-form": "^7.54.2",
-    "react-icons": "^5.5.0",
+    "ora": "^8.2.0",
+    "picocolors": "^1.1.1",
+    "prettier": "^3.6.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "rimraf": "^6.0.1",
-    "storybook": "^8.5.8",
-    "storybook-dark-mode": "^4.0.2",
-    "tslib": "^2.8.1",
-    "tsup": "^8.3.6",
-    "tsx": "^4.19.3",
-    "turbo": "^2.4.2",
-    "typescript": "^5.7.3",
-    "typescript-eslint": "^8.24.1",
-    "vite": "^6.3.4",
-    "vitest": "^3.1.2"
+    "storybook": "^9.0.13",
+    "tsdown": "^0.12.9",
+    "tsx": "^4.20.3",
+    "turbo": "^2.5.4",
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.35.0",
+    "vite": "^7.0.0",
+    "vitest": "^3.2.4",
+    "vitest-matchmedia-mock": "^1.0.6",
+    "web-features": "^2.40.1"
   },
   "pnpm": {
     "overrides": {
-      "react": "^18.2.0",
-      "react-dom": "^18.2.0"
+      "react": "^19.0.0",
+      "react-dom": "^19.0.0"
     }
   }
 }
