@@ -1,7 +1,8 @@
 {
   "name": "@yamada-ui/cli",
+  "type": "module",
   "version": "1.2.4",
-  "description": "Generate theme tokens for autocomplete",
+  "description": "The official CLI for Yamada UI projects",
   "keywords": [
     "theme",
     "theming",
@@ -36,27 +37,28 @@
     "yamada-cli": "bin/index.js"
   },
   "scripts": {
-    "clean": "rimraf dist .turbo",
-    "build": "tsup src --format cjs --clean --shims",
-    "dev": "pnpm build --watch",
-    "typecheck": " tsc --noEmit"
+    "clean": "rimraf dist .turbo .eslintcache",
+    "build": "tsdown",
+    "typecheck": " tsc --noEmit",
+    "lint": "eslint . --max-warnings=0 --cache",
+    "format": "prettier . --ignore-path ../../.prettierignore --check --cache"
   },
   "dependencies": {
-    "@clack/prompts": "^0.10.0",
+    "ora": "^8.2.0",
+    "picocolors": "^1.1.1",
     "chokidar": "^4.0.3",
     "cli-check-node": "^1.3.4",
     "cli-handle-unhandled": "^1.1.2",
-    "commander": "^13.1.0",
-    "esbuild": "^0.25.0",
-    "glob": "^11.0.1",
+    "commander": "^14.0.0",
+    "esbuild": "^0.25.5",
+    "glob": "^11.0.3",
     "node-eval": "^2.0.0",
-    "prettier": "^3.5.1",
+    "prettier": "^3.6.1",
     "tar": "^7.4.3",
-    "typescript": "^5.7.3"
+    "typescript": "^5.8.3"
   },
   "devDependencies": {
     "@types/update-notifier": "6.0.8",
-    "chalk": "^5.4.1",
     "update-notifier": "^7.3.1"
   }
 }
