@@ -36,140 +37,68 @@
     "url": "https://github.com/yamada-ui/yamada-ui/issues"
   },
   "scripts": {
-    "dev": "pnpm build:fast -- --watch",
-    "build": "tsup src --dts",
-    "build:fast": "tsup src",
-    "clean": "rimraf dist .turbo",
+    "clean": "rimraf dist .turbo .eslintcache",
+    "prebuild": "tsx scripts/pre-build.ts",
+    "build": "tsdown",
+    "postbuild": "tsx scripts/post-build.ts",
+    "storybook": "storybook dev -p 6006 -c storybook",
+    "storybook:serve": "npx http-server ../../storybook-static",
+    "storybook:build": "storybook build -c storybook -o ../../storybook-static",
+    "test": "vitest",
+    "test:dev": "vitest --coverage --watch --ui",
+    "test:ci": "vitest --coverage",
     "typecheck": "tsc --noEmit",
+    "lint": "eslint . --max-warnings=0 --cache",
+    "format": "prettier . --ignore-path ../../.prettierignore --check --cache",
     "prepack": "clean-package",
     "postpack": "clean-package restore"
   },
   "dependencies": {
-    "@yamada-ui/accordion": "workspace:*",
-    "@yamada-ui/alert": "workspace:*",
-    "@yamada-ui/autocomplete": "workspace:*",
-    "@yamada-ui/avatar": "workspace:*",
-    "@yamada-ui/badge": "workspace:*",
-    "@yamada-ui/breadcrumb": "workspace:*",
-    "@yamada-ui/button": "workspace:*",
-    "@yamada-ui/card": "workspace:*",
-    "@yamada-ui/checkbox": "workspace:*",
-    "@yamada-ui/close-button": "workspace:*",
-    "@yamada-ui/color-picker": "workspace:*",
-    "@yamada-ui/core": "workspace:*",
-    "@yamada-ui/data-list": "workspace:*",
-    "@yamada-ui/editable": "workspace:*",
-    "@yamada-ui/empty-state": "workspace:*",
-    "@yamada-ui/file-button": "workspace:*",
-    "@yamada-ui/file-input": "workspace:*",
-    "@yamada-ui/focus-lock": "workspace:*",
-    "@yamada-ui/for": "workspace:*",
-    "@yamada-ui/form-control": "workspace:*",
-    "@yamada-ui/format": "workspace:*",
-    "@yamada-ui/highlight": "workspace:*",
-    "@yamada-ui/icon": "workspace:*",
-    "@yamada-ui/image": "workspace:*",
-    "@yamada-ui/indicator": "workspace:*",
-    "@yamada-ui/infinite-scroll-area": "workspace:*",
-    "@yamada-ui/input": "workspace:*",
-    "@yamada-ui/kbd": "workspace:*",
-    "@yamada-ui/layouts": "workspace:*",
-    "@yamada-ui/link": "workspace:*",
-    "@yamada-ui/list": "workspace:*",
-    "@yamada-ui/loading": "workspace:*",
-    "@yamada-ui/menu": "workspace:*",
-    "@yamada-ui/modal": "workspace:*",
-    "@yamada-ui/motion": "workspace:*",
-    "@yamada-ui/native-select": "workspace:*",
-    "@yamada-ui/native-table": "workspace:*",
-    "@yamada-ui/notice": "workspace:*",
-    "@yamada-ui/number-input": "workspace:*",
-    "@yamada-ui/pagination": "workspace:*",
-    "@yamada-ui/password-input": "workspace:*",
-    "@yamada-ui/pin-input": "workspace:*",
-    "@yamada-ui/popover": "workspace:*",
-    "@yamada-ui/portal": "workspace:*",
-    "@yamada-ui/progress": "workspace:*",
-    "@yamada-ui/providers": "workspace:*",
-    "@yamada-ui/radio": "workspace:*",
-    "@yamada-ui/rating": "workspace:*",
-    "@yamada-ui/reorder": "workspace:*",
-    "@yamada-ui/resizable": "workspace:*",
-    "@yamada-ui/ripple": "workspace:*",
-    "@yamada-ui/scroll-area": "workspace:*",
-    "@yamada-ui/segmented-control": "workspace:*",
-    "@yamada-ui/select": "workspace:*",
-    "@yamada-ui/skeleton": "workspace:*",
-    "@yamada-ui/slider": "workspace:*",
-    "@yamada-ui/snacks": "workspace:*",
-    "@yamada-ui/stat": "workspace:*",
-    "@yamada-ui/status": "workspace:*",
-    "@yamada-ui/stepper": "workspace:*",
-    "@yamada-ui/switch": "workspace:*",
-    "@yamada-ui/tabs": "workspace:*",
-    "@yamada-ui/tag": "workspace:*",
-    "@yamada-ui/textarea": "workspace:*",
-    "@yamada-ui/theme": "workspace:*",
-    "@yamada-ui/theme-tools": "workspace:*",
-    "@yamada-ui/toggle": "workspace:*",
-    "@yamada-ui/tooltip": "workspace:*",
-    "@yamada-ui/transitions": "workspace:*",
-    "@yamada-ui/typography": "workspace:*",
-    "@yamada-ui/use-animation": "workspace:*",
-    "@yamada-ui/use-async-callback": "workspace:*",
-    "@yamada-ui/use-boolean": "workspace:*",
-    "@yamada-ui/use-breakpoint": "workspace:*",
-    "@yamada-ui/use-clickable": "workspace:*",
-    "@yamada-ui/use-clipboard": "workspace:*",
-    "@yamada-ui/use-controllable-state": "workspace:*",
-    "@yamada-ui/use-counter": "workspace:*",
-    "@yamada-ui/use-descendant": "workspace:*",
-    "@yamada-ui/use-disclosure": "workspace:*",
-    "@yamada-ui/use-event-listener": "workspace:*",
-    "@yamada-ui/use-eye-dropper": "workspace:*",
-    "@yamada-ui/use-focus": "workspace:*",
-    "@yamada-ui/use-focus-visible": "workspace:*",
-    "@yamada-ui/use-hover": "workspace:*",
-    "@yamada-ui/use-idle": "workspace:*",
-    "@yamada-ui/use-infinite-scroll": "workspace:*",
-    "@yamada-ui/use-interval": "workspace:*",
-    "@yamada-ui/use-latest-ref": "workspace:*",
-    "@yamada-ui/use-local-storage": "workspace:*",
-    "@yamada-ui/use-media-query": "workspace:*",
-    "@yamada-ui/use-os": "workspace:*",
-    "@yamada-ui/use-outside-click": "workspace:*",
-    "@yamada-ui/use-pan-event": "workspace:*",
-    "@yamada-ui/use-popper": "workspace:*",
-    "@yamada-ui/use-previous": "workspace:*",
-    "@yamada-ui/use-processing": "workspace:*",
-    "@yamada-ui/use-resize-observer": "workspace:*",
-    "@yamada-ui/use-size": "workspace:*",
-    "@yamada-ui/use-timeout": "workspace:*",
-    "@yamada-ui/use-token": "workspace:*",
-    "@yamada-ui/use-value": "workspace:*",
-    "@yamada-ui/use-window-event": "workspace:*",
+    "@emotion/css": "^11.13.5",
+    "@emotion/react": "^11.14.0",
+    "@emotion/serialize": "^1.3.3",
+    "@emotion/sheet": "^1.4.0",
+    "@emotion/styled": "^11.14.0",
+    "@emotion/use-insertion-effect-with-fallbacks": "^1.2.0",
+    "@emotion/utils": "^1.4.2",
+    "@floating-ui/react-dom": "^2.1.3",
+    "@tanstack/react-table": "^8.21.3",
     "@yamada-ui/utils": "workspace:*",
-    "@yamada-ui/visually-hidden": "workspace:*"
+    "color2k": "^2.0.3",
+    "copy-to-clipboard": "3.3.3",
+    "csstype": "^3.1.3",
+    "dayjs": "^1.11.13",
+    "embla-carousel": "^8.6.0",
+    "embla-carousel-react": "^8.6.0",
+    "framesync": "6.1.2",
+    "intl-messageformat": "^10.7.16",
+    "lucide-react": "0.523.0",
+    "motion": "12.18.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-dropzone-esm": "^15.2.0",
+    "react-fast-compare": "^3.2.2",
+    "react-focus-lock": "^2.13.6",
+    "react-remove-scroll": "^2.7.1",
+    "react-resizable-panels": "^3.0.2",
+    "recharts": "^3.0.0"
   },
   "devDependencies": {
+    "@faker-js/faker": "^9.8.0",
+    "@lucide/lab": "^0.1.2",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
     "clean-package": "2.2.0",
-    "react": "^18.3.1"
+    "react-hook-form": "^7.57.0",
+    "react-icons": "^5.5.0",
+    "use-mask-input": "^3.4.2"
   },
   "peerDependencies": {
     "@emotion/react": ">=11",
     "@emotion/styled": ">=11",
-    "motion": ">=11",
-    "react": ">=18",
-    "react-dom": ">=18"
+    "motion": ">=12",
+    "react": ">=19",
+    "react-dom": ">=19"
   },
-  "clean-package": "../../clean-package.config.json",
-  "tsup": {
-    "clean": true,
-    "target": "es2019",
-    "format": [
-      "cjs",
-      "esm"
-    ],
-    "sourcemap": true
-  }
+  "clean-package": "./clean-package.config.json"
 }
