@@ -1,9 +1,9 @@
-import type { Linter } from "eslint"
+import type { TSESLint } from "@typescript-eslint/utils"
 import cspellPlugin from "@cspell/eslint-plugin"
 import cspellJson from "../cspell.json"
 import { sharedFiles } from "./shared"
 
-export const cspellConfig: Linter.Config = {
+export const cspellConfig: TSESLint.FlatConfig.Config = {
   name: "eslint/cspell",
   files: sharedFiles,
   ignores: cspellJson.ignorePaths,
