@@ -1,16 +1,15 @@
-import type { Linter } from "eslint"
+import type { TSESLint } from "@typescript-eslint/utils"
 import { fixupPluginRules } from "@eslint/compat"
 import { flatConfigs } from "eslint-plugin-import"
-import importReplacePlugin from "eslint-plugin-import-replace"
 import unusedImportsPlugin from "eslint-plugin-unused-imports"
-import { sharedFiles, sharedTestFiles } from "./shared"
+import { sharedFiles } from "./shared"
 
-export const importConfigArray: Linter.Config[] = [
+export const importConfigArray: TSESLint.FlatConfig.ConfigArray = [
   {
     name: "eslint/import/order",
     files: sharedFiles,
     plugins: {
-      import: fixupPluginRules(flatConfigs.recommended.plugins.import),
+      import: fixupPluginRules(flatConfigs.recommended.plugins?.import ?? {}),
     },
     rules: {
       "import/consistent-type-specifier-style": ["error", "prefer-top-level"],
@@ -33,24 +32,6 @@ export const importConfigArray: Linter.Config[] = [
         node: true,
         typescript: true,
       },
-    },
-  },
-  {
-    name: "eslint/import/replace",
-    files: sharedTestFiles,
-    plugins: {
-      "import-replace": importReplacePlugin,
-    },
-    rules: {
-      "import-replace/import-replace": [
-        "error",
-        [
-          {
-            from: "@testing-library/react",
-            to: "@yamada-ui/test",
-          },
-        ],
-      ],
     },
   },
   {
