@@ -1,8 +1,8 @@
-import type { Linter } from "eslint"
+import type { TSESLint } from "@typescript-eslint/utils"
 import jsxA11yPlugin from "eslint-plugin-jsx-a11y"
 import { sharedFiles } from "./shared"
 
-export const jsxA11yConfig: Linter.Config = {
+export const jsxA11yConfig: TSESLint.FlatConfig.Config = {
   name: "eslint/jsx-a11y",
   files: sharedFiles,
   plugins: {
