@@ -5,45 +5,59 @@ import { sharedFiles } from "./shared"
 const type = "natural"
 
 const semanticSizes = {
-  base: "base",
-  "9xs": "9xs",
-  "8xs": "8xs",
-  "7xs": "7xs",
-  "6xs": "6xs",
-  "5xs": "5xs",
-  "4xs": "4xs",
-  "3xs": "3xs",
-  "2xs": ["2xs", "hairline", "ultra-fast"],
-  xs: ["xs", "thin", "tighter", "faster", "shorter"],
-  sm: ["sm", "light", "tight", "fast", "short"],
-  md: ["md", "normal"],
-  lg: ["lg", "medium", "wide", "slow", "tall"],
-  xl: ["xl", "semibold", "wider", "slower", "taller"],
-  "2xl": ["2xl", "bold", "widest", "ultra-slow"],
-  "3xl": ["3xl", "extrabold"],
-  "4xl": ["4xl", "black"],
-  "5xl": "5xl",
-  "6xl": "6xl",
-  "7xl": "7xl",
-  "8xl": "8xl",
-  "9xl": "9xl",
+  base: "^base$",
+  "9xs": "^9xs$",
+  "8xs": "^8xs$",
+  "7xs": "^7xs$",
+  "6xs": "^6xs$",
+  "5xs": "^5xs$",
+  "4xs": "^4xs$",
+  "3xs": "^3xs$",
+  "2xs": ["^2xs$", "^hairline$", "^fastest$"],
+  xs: ["^xs$", "^thin$", "^tighter$", "^faster$", "^shorter$"],
+  sm: ["^sm$", "^light$", "^tight$", "^fast$", "^short$"],
+  md: ["^md$", "^normal$", "^moderate$"],
+  lg: ["^lg$", "^medium$", "^wide$", "^slow$", "^tall$"],
+  xl: ["^xl$", "^semibold$", "^wider$", "^slower$", "^taller$"],
+  "2xl": ["^2xl$", "^bold$", "^widest$", "^slowest$"],
+  "3xl": ["^3xl$", "^extrabold$"],
+  "4xl": ["^4xl$", "^black$"],
+  "5xl": "^5xl$",
+  "6xl": "^6xl$",
+  "7xl": "^7xl$",
+  "8xl": "^8xl$",
+  "9xl": "^9xl$",
 }
+
+const sortPseudos = [
+  ["^_hover$"],
+  ["^_active$", "^_focus$", "^_focusVisible$"],
+  ["^_invalid$", "^_readOnly$", "^_hidden$"],
+  ["^_disabled$", "^_never$"],
+]
+
+const otherPseudos = sortPseudos
+  .flatMap((pseudos) => pseudos.map((pseudo) => pseudo.replace(/\^|_|\$/g, "")))
+  .join("|")
 
 const sortObjectGroups = {
   customGroups: {
-    aria: "aria-*",
-    callback: "on*",
-    data: "data-*",
-    internal: "__*",
-    primary: ["key", "ref", "id", "lang"],
-    props: "*Props",
-    pseudos: "_*",
-    quaternary: ["className", "alt"],
-    quinary: ["css", "sx", "style"],
-    secondary: ["as", "form", "type", "htmlFor"],
-    senary: ["layerStyle", "textStyle", "baseStyle", "apply"],
-    septenary: ["variant", "size", "colorScheme"],
-    tertiary: ["name", "src", "srcSet", "href", "target"],
+    aria: "^aria-.+",
+    callback: "^on.*",
+    data: "^data-.+",
+    internal: "^__.+",
+    primary: ["^key$", "^ref$", "^id$", "^lang$"],
+    props: ".+Props$",
+    pseudos: `^_(?!.*(${otherPseudos})$).*$`,
+    ...Object.fromEntries(
+      sortPseudos.map((pseudos, index) => [`pseudos${index + 1}`, pseudos]),
+    ),
+    quaternary: ["^className$", "^alt$"],
+    quinary: ["^css$", "^sx$", "^style$"],
+    secondary: ["^as$", "^form$", "^type$", "^htmlFor$"],
+    senary: ["^layerStyle$", "^textStyle$", "^apply$"],
+    septenary: ["^variant$", "^size$", "^colorScheme$"],
+    tertiary: ["^name$", "^src$", "^srcSet$", "^href$", "^target$", "^rel$"],
     ...semanticSizes,
   },
   groups: [
@@ -58,6 +72,7 @@ const sortObjectGroups = {
     ["aria", "data"],
     "unknown",
     "pseudos",
+    ...sortPseudos.map((_, index) => `pseudos${index + 1}`),
     "props",
     "callback",
     "internal",
@@ -69,7 +84,20 @@ export const perfectionistConfig: TSESLint.FlatConfig.Config = {
   files: sharedFiles,
   plugins: { perfectionist: perfectionistPlugin },
   rules: {
-    "perfectionist/sort-exports": ["error", { type }],
+    "perfectionist/sort-exports": [
+      "error",
+      {
+        type,
+        customGroups: [
+          {
+            anyOf: [{ elementNamePattern: [".style(.js|.jsx|.ts|.tsx)?$"] }],
+            groupName: "style",
+          },
+        ],
+        groups: ["style"],
+        partitionByNewLine: true,
+      },
+    ],
     "perfectionist/sort-imports": [
       "error",
       {
@@ -85,7 +113,8 @@ export const perfectionistConfig: TSESLint.FlatConfig.Config = {
           "unknown",
           ["side-effect", "side-effect-style"],
         ],
-        newlinesBetween: "never",
+        newlinesBetween: "ignore",
+        partitionByNewLine: true,
       },
     ],
 
