@@ -1,2 +1,2 @@
-export { actionTheme } from "./theme"
-export { actionTokens, themePath } from "./tokens"
+export { actionTheme } from "./theme/index.js"
+export { actionTokens } from "./tokens/index.js"
