@@ -1,107 +1,101 @@
-import type { Dict } from "../../utils"
-import * as p from "@clack/prompts"
-import c from "chalk"
+import type { Dict } from "../../utils/index.js"
 import chokidar from "chokidar"
 import { writeFile } from "fs/promises"
+import ora from "ora"
 import path from "path"
-import { getModule } from "../../utils"
-import { createThemeTypings } from "./create-theme-typings"
-import { resolveOutputPath, themePath } from "./resolve-output-path"
+import c from "picocolors"
+import { getModule, isString } from "../../utils/index.js"
+import { generateThemeTokens } from "./generate-theme-tokens.js"
 
-const generateThemeTypings = async ({
+const buildThemeTokens = async ({
   config,
-  outFile,
+  internal,
+  outPath,
   theme,
 }: {
   config: Dict
+  internal: boolean
+  outPath: string
   theme: Dict
-  outFile?: string
 }) => {
-  p.intro(c.magenta(`Generating Yamada UI theme typings`))
+  const themeTokens = await generateThemeTokens(
+    theme,
+    config.theme ?? {},
+    internal,
+  )
 
-  const s = p.spinner()
-
-  try {
-    const start = process.hrtime.bigint()
-
-    s.start(`Parsing the theme`)
-
-    const generatedTheme = await createThemeTypings(theme, config.theme ?? {})
-
-    s.stop(`Parsed the theme`)
-
-    s.start(`Resolving the output path`)
-
-    const outPath = await resolveOutputPath(outFile)
-
-    s.stop(`Resolved the output path`)
-
-    s.start(`Writing file "${outPath}"`)
-
-    await writeFile(outPath, generatedTheme, "utf8")
-
-    s.stop(`Wrote file`)
-
-    p.note(outPath, "Output path")
-
-    const end = process.hrtime.bigint()
-    const duration = (Number(end - start) / 1e9).toFixed(2)
-
-    p.outro(`${c.green(`Done`)} in ${c.dim(`${duration}s`)}\n`)
-  } catch (e) {
-    s.stop(`An error occurred`, 500)
-
-    p.cancel(c.red(e instanceof Error ? e.message : "Message is missing"))
-  }
+  await writeFile(outPath, themeTokens, "utf8")
 }
 
-export { themePath }
+const getTheme = async (path: string, cwd: string) => {
+  const { dependencies, mod } = await getModule(path, cwd)
+
+  const theme =
+    mod?.default ?? mod?.theme ?? mod?.customTheme ?? mod?.defaultTheme ?? {}
+  const config = mod?.config ?? mod?.customConfig ?? mod?.defaultConfig ?? {}
+
+  return { config, dependencies, theme }
+}
 
 interface Options {
   cwd?: string
+  internal?: boolean
   out?: string
-  watch?: string
+  watch?: boolean | string
 }
 
 export const actionTokens = async (
-  themePath: string,
-  { cwd = path.resolve(), out: outFile, watch: watchFile }: Options,
+  inputPath: string,
+  { cwd = path.resolve(), internal = false, out: outPath, watch }: Options,
 ) => {
-  const readFile = async () => {
-    const filePath = path.resolve(themePath)
-    const { dependencies, mod } = await getModule(filePath, cwd)
+  const spinner = ora()
 
-    const theme =
-      mod?.default ?? mod?.theme ?? mod?.customTheme ?? mod?.defaultTheme ?? {}
-    const config = mod?.config ?? mod?.customConfig ?? mod?.defaultConfig ?? {}
+  spinner.start(`Getting theme`)
 
-    return { config, dependencies, theme }
-  }
+  inputPath = path.resolve(cwd, inputPath)
 
-  let file = await readFile()
+  let file = await getTheme(inputPath, cwd)
 
   const { config, dependencies, theme } = file
 
-  const buildFile = async () => {
-    await generateThemeTypings({ config, outFile, theme })
+  spinner.succeed(`Got theme`)
 
-    if (watchFile) console.log("\n", "⌛️ Watching for changes...")
+  const buildFile = async () => {
+    spinner.start(`Generating theme typings`)
+
+    await buildThemeTokens({
+      config,
+      internal,
+      outPath: outPath ?? path.join(inputPath, "index.types.ts"),
+      theme,
+    })
+
+    spinner.succeed(`Generated theme typings`)
+
+    if (watch) spinner.info("Watching for changes...")
   }
 
-  if (watchFile) {
-    const watchPath = typeof watchFile === "string" ? watchFile : dependencies
+  if (watch) {
+    const watchPath = isString(watch) ? watch : dependencies
 
     chokidar
       .watch(watchPath)
       .on("ready", buildFile)
-      .on("change", async (filePath) => {
-        console.log("📦 File changed", filePath)
+      .on("change", async (path) => {
+        spinner.info(`File changed ${path}`)
 
-        file = await readFile()
+        file = await getTheme(inputPath, cwd)
 
         return buildFile()
       })
   } else {
+    const start = process.hrtime.bigint()
+
     await buildFile()
+
+    const end = process.hrtime.bigint()
+    const duration = (Number(end - start) / 1e9).toFixed(2)
+
+    console.log("\n", c.green(`Done in ${duration}s`))
   }
 }
