@@ -1,6 +1,6 @@
-export * from "./assertion"
-export * from "./cli"
-export * from "./fs"
-export * from "./module"
-export * from "./object"
-export * from "./prettier"
+export * from "./assertion.js"
+export * from "./cli.js"
+export * from "./fs.js"
+export * from "./module.js"
+export * from "./object.js"
+export * from "./prettier.js"
