@@ -2,7 +2,7 @@ import { build } from "esbuild"
 import { realpathSync } from "fs"
 import nodeEval from "node-eval"
 import { Script } from "vm"
-import { isObject } from "./assertion"
+import { isObject } from "./assertion.js"
 
 export const getModule = async (file: string, cwd: string) => {
   const result = await build({
