@@ -1,4 +1,10 @@
-import type { Dict } from "./assertion"
+import type { Dict } from "./assertion.js"
+
+export const getObject = (obj: Dict, path: string) => {
+  const keys = path.split(".")
+
+  return keys.reduce((obj, key) => obj[key] ?? {}, obj)
+}
 
 export const omitObject = <T extends Dict, K extends keyof T>(
   obj: T,
@@ -13,10 +19,4 @@ export const omitObject = <T extends Dict, K extends keyof T>(
   })
 
   return result as Omit<T, K>
-}
-
-export const getObject = (obj: Dict, path: string) => {
-  const keys = path.split(".")
-
-  return keys.reduce((obj, key) => obj[key] ?? {}, obj)
 }
