@@ -1,27 +1,22 @@
-import type { Dict, StringLiteral } from "./index.types"
 import * as c from "color2k"
-import { isArray, isNumber, isObject } from "./assertion"
-import { getMemoizedObject as get } from "./object"
-
-type ThemeValue = number | string
-type Breakpoint = "base" | StringLiteral
-type ColorMode = "dark" | "light"
+import { isNumber } from "./assertion"
 
 export type ColorFormat = "hex" | "hexa" | "hsl" | "hsla" | "rgb" | "rgba"
 
 export const SEMANTIC_COLOR_SCHEMES = [
+  "mono",
   "primary",
   "secondary",
   "info",
   "success",
   "warning",
   "danger",
+  "error",
   "link",
 ] as const
 
-export const COLOR_SCHEMES = [
+export const DEFAULT_COLOR_SCHEMES = [
   "gray",
-  "neutral",
   "red",
   "rose",
   "pink",
@@ -42,228 +37,14 @@ export const COLOR_SCHEMES = [
   "fuchsia",
 ] as const
 
-export type SemanticColorScheme =
-  | (typeof SEMANTIC_COLOR_SCHEMES)[number]
-  | StringLiteral
-export type ColorScheme = (typeof COLOR_SCHEMES)[number] | StringLiteral
+export const COLOR_SCHEMES = [
+  ...SEMANTIC_COLOR_SCHEMES,
+  ...DEFAULT_COLOR_SCHEMES,
+]
 
 export const TONES = [
   50, 100, 200, 300, 400, 500, 600, 700, 800, 900, 950,
 ] as const
-
-export function isGray(colorScheme: string) {
-  return colorScheme === "gray" || colorScheme === "neutral"
-}
-
-export function isAccessible(colorScheme: string) {
-  return (
-    colorScheme === "yellow" || colorScheme === "cyan" || colorScheme === "lime"
-  )
-}
-
-export function getColor(color: string, fallback = "#000000") {
-  return function (
-    theme: Dict = {},
-    colorMode: ColorMode = "light",
-    breakpoint: Breakpoint = "base",
-  ) {
-    const [token, tone] = color.split(".")
-
-    if (tone) {
-      const [, relatedToken] =
-        Object.entries<string>(theme.semantics?.colorSchemes ?? {}).find(
-          ([semanticToken]) => token === semanticToken,
-        ) ?? []
-
-      if (relatedToken) color = `${relatedToken}.${tone}`
-    } else {
-      const [, relatedColor] =
-        Object.entries<string>(theme.semantics?.colors ?? {}).find(
-          ([semanticToken]) => token === semanticToken,
-        ) ?? []
-
-      if (relatedColor) color = relatedColor
-    }
-
-    const hex = get<[ThemeValue, ThemeValue] | Dict<ThemeValue> | ThemeValue>(
-      theme,
-      `colors.${color}`,
-      color,
-    )
-
-    try {
-      if (isArray(hex)) {
-        return c.toHex(String(hex[colorMode !== "dark" ? 0 : 1]))
-      } else if (isObject(hex)) {
-        return c.toHex(String(hex[breakpoint]))
-      } else {
-        return c.toHex(String(hex))
-      }
-    } catch {
-      try {
-        return c.toHex(fallback)
-      } catch {
-        return "#000000"
-      }
-    }
-  }
-}
-
-export function lightenColor(color: string, amount: number) {
-  return function (
-    theme?: Dict,
-    colorMode?: ColorMode,
-    breakpoint?: Breakpoint,
-  ) {
-    const raw = getColor(color, color)(theme, colorMode, breakpoint)
-
-    return c.toHex(c.lighten(raw, amount / 100))
-  }
-}
-
-export function darkenColor(color: string, amount: number) {
-  return function (
-    theme?: Dict,
-    colorMode?: ColorMode,
-    breakpoint?: Breakpoint,
-  ) {
-    const raw = getColor(color, color)(theme, colorMode, breakpoint)
-
-    return c.toHex(c.darken(raw, amount / 100))
-  }
-}
-
-export function tintColor(color: string, amount: number) {
-  return function (
-    theme?: Dict,
-    colorMode?: ColorMode,
-    breakpoint?: Breakpoint,
-  ) {
-    const raw = getColor(color, color)(theme, colorMode, breakpoint)
-
-    return c.toHex(c.mix(raw, "#fff", amount / 100))
-  }
-}
-
-export function shadeColor(color: string, amount: number) {
-  return function (
-    theme?: Dict,
-    colorMode?: ColorMode,
-    breakpoint?: Breakpoint,
-  ) {
-    const raw = getColor(color, color)(theme, colorMode, breakpoint)
-
-    return c.toHex(c.mix(raw, "#000", amount / 100))
-  }
-}
-
-export function transparentizeColor(color: string, alpha: number) {
-  return function (
-    theme?: Dict,
-    colorMode?: ColorMode,
-    breakpoint?: Breakpoint,
-  ) {
-    const raw = getColor(color, color)(theme, colorMode, breakpoint)
-
-    return c.transparentize(raw, 1 - alpha)
-  }
-}
-
-export function randomColor({
-  colors,
-  string,
-}: { colors?: string[]; string?: string } = {}) {
-  const fallback = randomHex()
-
-  if (string && colors) return randomColorFromList(string, colors) ?? fallback
-
-  if (string && !colors) return randomColorFromString(string)
-
-  if (colors && !string) return randomFromList(colors) ?? fallback
-
-  return fallback
-}
-
-function randomHex() {
-  return `#${Math.floor(Math.random() * 0xffffff)
-    .toString(16)
-    .padEnd(6, "0")}`
-}
-
-function randomColorFromString(str: string) {
-  let hash = 0
-
-  if (str.length === 0) return hash.toString()
-
-  for (let i = 0; i < str.length; i += 1) {
-    hash = str.charCodeAt(i) + ((hash << 5) - hash)
-    hash = hash & hash
-  }
-
-  let color = "#"
-
-  for (let j = 0; j < 3; j += 1) {
-    const value = (hash >> (j * 8)) & 255
-
-    color += `00${value.toString(16)}`.substr(-2)
-  }
-
-  return color
-}
-
-function randomColorFromList(str: string, list: string[]) {
-  let index = 0
-
-  if (str.length === 0) return list[0]
-
-  for (let i = 0; i < str.length; i += 1) {
-    index = str.charCodeAt(i) + ((index << 5) - index)
-
-    index = index & index
-  }
-
-  index = ((index % list.length) + list.length) % list.length
-
-  return list[index]
-}
-
-function randomFromList(list: string[]) {
-  return list[Math.floor(Math.random() * list.length)]
-}
-
-function getBrightness(color: string) {
-  const [r, g, b] = c.parseToRgba(color)
-
-  return (r * 299 + g * 587 + b * 114) / 1000
-}
-
-export function isTone(color: string) {
-  return function (
-    theme?: Dict,
-    colorMode?: ColorMode,
-    breakpoint?: Breakpoint,
-  ) {
-    const raw = theme ? getColor(color)(theme, colorMode, breakpoint) : color
-
-    const brightness = getBrightness(raw)
-
-    const isDark = brightness < 128
-
-    return isDark ? "dark" : "light"
-  }
-}
-
-export function isLight(color: string) {
-  return function (theme?: Dict, colorMode?: ColorMode) {
-    return isTone(color)(theme, colorMode) === "dark"
-  }
-}
-
-export function isDark(color: string) {
-  return function (theme?: Dict, colorMode?: ColorMode) {
-    return isTone(color)(theme, colorMode) === "light"
-  }
-}
 
 export function convertColor(color: string, fallback?: string) {
   return function (format: ColorFormat): string | undefined {
