@@ -1,5 +1,5 @@
-import type React from "react"
-import { isNumber, isUndefined } from "./assertion"
+import { isArray, isNumber, isObject, isUndefined } from "./assertion"
+import { noop } from "./function"
 
 export function createdDom(): boolean {
   return !!(
@@ -13,6 +13,21 @@ export function getPlatform(): string {
   return (navigator as any).userAgentData?.platform ?? navigator.platform
 }
 
+export function getUserAgent(): string {
+  const userAgentData = (navigator as any).userAgentData
+
+  if (userAgentData && isArray(userAgentData.brands)) {
+    return userAgentData.brands
+      .map(
+        ({ brand, version }: { brand: string; version: string }) =>
+          `${brand}/${version}`,
+      )
+      .join(" ")
+  }
+
+  return navigator.userAgent
+}
+
 export function vendor(v: RegExp): boolean {
   return createdDom() && v.test(navigator.vendor)
 }
@@ -21,42 +36,94 @@ export function platform(v: RegExp): boolean {
   return createdDom() && v.test(getPlatform())
 }
 
+export function userAgent(v: RegExp): boolean {
+  return createdDom() && v.test(getUserAgent())
+}
+
 export function isMac(): boolean {
-  return platform(/^mac/i)
+  return platform(/^Mac/i)
+}
+
+export function isIPhone(): boolean {
+  return platform(/^iPhone/i)
+}
+
+export function isIPad(): boolean {
+  return platform(/^iPad/i) || (isMac() && navigator.maxTouchPoints > 1)
+}
+
+export function isIos(): boolean {
+  return isIPhone() || isIPad()
+}
+
+export function isAndroid(): boolean {
+  return platform(/^Android/i)
 }
 
 export function isApple(): boolean {
-  return platform(/mac|iphone|ipad|ipod/i)
+  return isMac() || isIos()
 }
 
 export function isSafari(): boolean {
   return isApple() && vendor(/apple/i)
 }
 
-export function isElement(el: any): el is Element {
-  return (
-    el != null &&
-    typeof el == "object" &&
-    "nodeType" in el &&
-    el.nodeType === Node.ELEMENT_NODE
-  )
+export function isFirefox(): boolean {
+  return userAgent(/Firefox/i)
+}
+
+export function isChrome(): boolean {
+  return userAgent(/Chrome/i)
+}
+
+export function isWebKit(): boolean {
+  return userAgent(/AppleWebKit/i) && !isChrome()
 }
 
 export function isHTMLElement(el: any): el is HTMLElement {
-  if (!isElement(el)) return false
-
-  const win = el.ownerDocument.defaultView ?? window
-
-  return el instanceof win.HTMLElement
+  return isNode(el) && el.nodeType === Node.ELEMENT_NODE
 }
 
-export function isHidden(el: HTMLElement): boolean {
-  if (el.parentElement && isHidden(el.parentElement)) return true
+export function isWindow(el: any): el is Window {
+  return isObject(el) && el === el.window
+}
+
+export function isDocument(el: any): el is Document {
+  return isNode(el) && el.nodeType === Node.DOCUMENT_NODE
+}
+
+export function isVisualViewport(el: any): el is VisualViewport {
+  return isObject(el) && el.constructor.name === "VisualViewport"
+}
+
+export function isNode(el: any): el is Node {
+  return isObject(el) && el.nodeType !== undefined
+}
+
+export function isShadowRoot(el: any): el is ShadowRoot {
+  return (
+    isNode(el) && el.nodeType === Node.DOCUMENT_FRAGMENT_NODE && "host" in el
+  )
+}
+
+export function isFrame(el: any): el is HTMLIFrameElement {
+  return isHTMLElement(el) && el.tagName === "IFRAME"
+}
+
+export function isActiveElement(
+  el: HTMLElement,
+  rootNode: Document | ShadowRoot,
+): boolean {
+  return getActiveElement(rootNode) === el
+}
+
+export function isHiddenElement(el: HTMLElement): boolean {
+  if (el.parentElement && isHiddenElement(el.parentElement)) return true
 
   return el.hidden
 }
 
-export function isDisabled(el: HTMLElement): boolean {
+export function isDisabledElement(el: HTMLElement): boolean {
   return (
     Boolean(el.getAttribute("disabled")) ||
     Boolean(el.getAttribute("data-disabled")) ||
@@ -64,25 +131,209 @@ export function isDisabled(el: HTMLElement): boolean {
   )
 }
 
-function isVisible(el: HTMLElement) {
-  return el.offsetWidth > 0 && el.offsetHeight > 0
+export function isVisibleElement(el?: Node): boolean {
+  if (!isHTMLElement(el)) return false
+
+  return (
+    el.offsetWidth > 0 || el.offsetHeight > 0 || el.getClientRects().length > 0
+  )
 }
 
-export function hasTabIndex(el: HTMLElement): boolean {
-  return el.hasAttribute("tabindex")
+export function isTouchDevice(): boolean {
+  return "ontouchstart" in window
 }
 
-export function isContentEditable(el: HTMLElement): boolean {
-  const value = el.getAttribute("contenteditable")
+const hasTabIndex = (el: Element): boolean =>
+  !Number.isNaN(parseInt(el.getAttribute("tabindex") || "0", 10))
 
-  return value !== "false" && value != null
+const hasNegativeTabIndex = (el: Element): boolean =>
+  parseInt(el.getAttribute("tabindex") || "0", 10) < 0
+
+export function getWindow(
+  el: Document | Node | null | ShadowRoot | undefined,
+): typeof globalThis & Window {
+  if (isShadowRoot(el)) return getWindow(el.host)
+  if (isDocument(el)) return el.defaultView ?? window
+  if (isHTMLElement(el)) return el.ownerDocument.defaultView ?? window
+
+  return window
 }
 
-export function isContains(
-  parent: HTMLElement | null,
-  child: HTMLElement | null,
-): boolean | undefined {
-  return parent === child || parent?.contains(child)
+export function getDocument(
+  el: Document | Element | Node | null | undefined | Window,
+): Document {
+  if (isDocument(el)) return el
+  if (isWindow(el)) return el.document
+
+  return el?.ownerDocument ?? document
+}
+
+export function getDocumentElement(
+  el: Document | Element | Node | null | undefined | Window,
+): HTMLElement {
+  return getDocument(el).documentElement
+}
+
+export function getNodeName(node: Node | Window): string {
+  if (isHTMLElement(node)) return node.localName || ""
+
+  return "#document"
+}
+
+export function getActiveElement(
+  rootNode: Document | ShadowRoot,
+): HTMLElement | null {
+  let activeEl = rootNode.activeElement as HTMLElement | null
+
+  while (activeEl?.shadowRoot) {
+    const el = activeEl.shadowRoot.activeElement as HTMLElement | null
+
+    if (el === activeEl) {
+      break
+    } else {
+      activeEl = el
+    }
+  }
+
+  return activeEl
+}
+
+export function getTabbableElements(
+  el: HTMLElement | null,
+  includeEl?: boolean,
+) {
+  if (!el) return []
+  const els = Array.from(el.querySelectorAll<HTMLElement>(focusableSelector))
+  const tabbableEls = els.filter(isTabbableElement)
+
+  if (includeEl && isTabbableElement(el)) {
+    tabbableEls.unshift(el)
+  }
+
+  tabbableEls.forEach((el, i) => {
+    if (isFrame(el) && el.contentDocument) {
+      const frameBody = el.contentDocument.body
+      const allFrameTabbable = getTabbableElements(frameBody)
+      tabbableEls.splice(i, 1, ...allFrameTabbable)
+    }
+  })
+
+  if (!tabbableEls.length && includeEl) {
+    return els
+  }
+
+  return tabbableEls
+}
+
+export function getFirstTabbableElement(
+  el: HTMLElement | null,
+  includeEl?: boolean,
+): HTMLElement | null {
+  const [first] = getTabbableElements(el, includeEl)
+
+  return first || null
+}
+
+export function getNextTabbableElement(
+  el: HTMLElement | null,
+  current?: HTMLElement | null,
+): HTMLElement | null {
+  const els = getTabbableElements(el)
+  const doc = el?.ownerDocument || document
+  const currentElement = current ?? (doc.activeElement as HTMLElement | null)
+
+  if (!currentElement) return null
+
+  return els[els.indexOf(currentElement) + 1] || null
+}
+
+export function getLastTabbableElement(
+  el: HTMLElement | null,
+  includeEl?: boolean,
+): HTMLElement | null {
+  const els = getTabbableElements(el, includeEl)
+
+  return els[els.length - 1] || null
+}
+
+export function getTabbableElementEdges(
+  el: HTMLElement | null,
+  includeEl?: boolean,
+): [HTMLElement | null, HTMLElement | null] {
+  const els = getTabbableElements(el, includeEl)
+  const first = els[0] || null
+  const last = els[els.length - 1] || null
+
+  return [first, last]
+}
+
+export function getParentNode(node: Node): Node {
+  if (getNodeName(node) === "html") return node
+
+  const result =
+    (node as any).assignedSlot ||
+    node.parentNode ||
+    isShadowRoot(node) ||
+    getDocumentElement(node)
+
+  return isShadowRoot(result) ? result.host : result
+}
+
+export function isInputElement(el: any): el is HTMLInputElement {
+  return isHTMLElement(el) && el.localName === "input"
+}
+
+export function isEditableElement(
+  el: EventTarget | HTMLElement | null,
+): boolean {
+  if (el == null || !isHTMLElement(el)) return false
+  try {
+    return (
+      (isInputElement(el) && el.selectionStart != null) ||
+      /(textarea|select)/.test(el.localName) ||
+      el.isContentEditable ||
+      el.getAttribute("contenteditable") === "true" ||
+      el.getAttribute("contenteditable") === ""
+    )
+  } catch {
+    return false
+  }
+}
+
+export function isFocusableElement(el: HTMLElement | null): el is HTMLElement {
+  if (!el || el.closest("[inert]")) return false
+
+  return el.matches(focusableSelector) && isVisibleElement(el)
+}
+
+export function isTabbableElement(el: HTMLElement | null): el is HTMLElement {
+  if (el != null && el.tabIndex > 0) return true
+  return isFocusableElement(el) && !hasNegativeTabIndex(el as Element)
+}
+
+export function contains(
+  parent?: EventTarget | HTMLElement | null,
+  child?: EventTarget | HTMLElement | null,
+): boolean {
+  if (!parent || !child) return false
+  if (!isHTMLElement(parent) || !isHTMLElement(child)) return false
+
+  const rootNode = child.getRootNode()
+
+  if (parent === child) return true
+  if (parent.contains(child)) return true
+
+  if (isShadowRoot(rootNode)) {
+    let next: Node | null | ShadowRoot = child
+
+    while (next) {
+      if (parent === next) return true
+
+      next = next.parentNode ?? (isShadowRoot(next) ? next.host : null)
+    }
+  }
+
+  return false
 }
 
 export function getPx(value: number | string | undefined): number {
@@ -105,105 +356,186 @@ export function getPx(value: number | string | undefined): number {
   return parseFloat(value) * fontSize
 }
 
-export function getEventRelatedTarget(ev: React.FocusEvent | React.MouseEvent) {
-  return (ev.relatedTarget ??
-    ev.currentTarget.ownerDocument.activeElement) as HTMLElement | null
+export function getTabIndex(node: HTMLElement | SVGElement) {
+  if (node.tabIndex < 0) {
+    if (
+      (/^(audio|video|details)$/.test(node.localName) ||
+        isEditableElement(node)) &&
+      !hasTabIndex(node)
+    ) {
+      return 0
+    }
+  }
+  return node.tabIndex
+}
+
+export function dataAttr(condition: any): string | undefined {
+  return (condition ? "" : undefined) as string | undefined
 }
 
 type Booleanish = "false" | "true" | boolean
 
-export function dataAttr(condition: boolean | undefined) {
-  return (condition ? "" : undefined) as Booleanish
+export function ariaAttr(condition: any): Booleanish | undefined {
+  if (condition === "true") return "true"
+  if (condition === "false") return "false"
+
+  return !!condition ? true : undefined
 }
 
-export function ariaAttr(condition: boolean | undefined): boolean | undefined {
-  return condition ? true : undefined
-}
-
-export interface FocusableElement {
-  focus: (options?: FocusOptions) => void
+export function uuid() {
+  return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx"
+    .replace(/x/g, () => Math.random().toString(16).slice(2, 3))
+    .replace(/y/g, () => ((Math.random() * 4) | 0).toString(16))
 }
 
 const focusableElList = [
-  "input:not(:disabled):not([disabled])",
-  "select:not(:disabled):not([disabled])",
-  "textarea:not(:disabled):not([disabled])",
-  "embed",
+  "[tabindex]",
+  "input:not([type='hidden']):not([disabled])",
+  "select:not([disabled])",
+  "textarea:not([disabled])",
+  "button:not([disabled])",
   "iframe",
   "object",
+  "embed",
   "a[href]",
   "area[href]",
-  "button:not(:disabled):not([disabled])",
-  "[tabindex]",
   "audio[controls]",
   "video[controls]",
-  "*[tabindex]:not([aria-disabled])",
-  "*[contenteditable]",
+  "[contenteditable]:not([contenteditable='false'])",
+  "details > summary:first-of-type",
 ]
 
-const focusableElSelector: string = focusableElList.join()
+const focusableSelector = focusableElList.join(", ")
 
-export function getAllFocusable<T extends HTMLElement>(container: T): T[] {
-  const focusableEls: T[] = Array.from(
-    container.querySelectorAll<T>(focusableElSelector),
+export function getFocusableElements(
+  el: null | Pick<HTMLElement, "querySelectorAll">,
+  includeEl = false,
+) {
+  if (!el) return []
+  const els = Array.from(el.querySelectorAll<HTMLElement>(focusableSelector))
+
+  const include = includeEl
+
+  if (include && isHTMLElement(el) && isFocusableElement(el)) {
+    els.unshift(el)
+  }
+
+  const focusableEls = els.filter(isFocusableElement)
+
+  focusableEls.forEach((el, i) => {
+    if (isFrame(el) && el.contentDocument) {
+      const frameBody = el.contentDocument.body
+
+      focusableEls.splice(i, 1, ...getFocusableElements(frameBody))
+    }
+  })
+
+  return focusableEls
+}
+
+export function getFirstFocusableElement(
+  el: HTMLElement | null,
+  includeEl?: boolean,
+): HTMLElement | null {
+  const [first] = getFocusableElements(el, includeEl)
+
+  return first || null
+}
+
+export function setAttribute(el: Element, attr: string, v: string) {
+  const prev = el.getAttribute(attr)
+  const exists = prev != null
+
+  el.setAttribute(attr, v)
+
+  return () => {
+    if (!exists) {
+      el.removeAttribute(attr)
+    } else {
+      el.setAttribute(attr, prev)
+    }
+  }
+}
+
+export function setStyle(
+  el: HTMLElement | null | undefined,
+  style: Partial<
+    CSSStyleDeclaration | { [key: `--${string}`]: number | string }
+  >,
+) {
+  if (!el) return noop
+
+  const prev = Object.keys(style).reduce<{ [key: string]: string }>(
+    (acc, key) => {
+      acc[key] = el.style.getPropertyValue(key)
+
+      return acc
+    },
+    {},
   )
 
-  focusableEls.unshift(container)
+  Object.assign(el.style, style)
 
-  return focusableEls.filter((el) => isFocusable(el) && isVisible(el))
+  return () => {
+    Object.assign(el.style, prev)
+
+    if (el.style.length === 0) el.removeAttribute("style")
+  }
 }
 
-export function isFocusable(el: HTMLElement): boolean {
-  if (!isHTMLElement(el) || isHidden(el) || isDisabled(el)) {
-    return false
+const LOCK_DATA_ATTR = "data-scroll-lock"
+
+export function scrollLock(
+  el: Document | Element | Node | null | undefined | Window,
+) {
+  const doc = getDocument(el)
+  const win = getWindow(doc)
+
+  const locked = doc.body.hasAttribute(LOCK_DATA_ATTR)
+
+  if (locked) return
+
+  const scrollbarWidth = win.innerWidth - doc.documentElement.clientWidth
+  const documentLeft = doc.documentElement.getBoundingClientRect().left
+  const scrollbarX = Math.round(documentLeft) + doc.documentElement.scrollLeft
+  const paddingProperty = scrollbarX ? "paddingLeft" : "paddingRight"
+
+  doc.body.setAttribute(LOCK_DATA_ATTR, "")
+
+  const setStyleForIOS = () => {
+    const offsetLeft = win.visualViewport?.offsetLeft ?? 0
+    const offsetTop = win.visualViewport?.offsetTop ?? 0
+
+    const restoreStyle = setStyle(doc.body, {
+      left: `${-(win.scrollX - Math.floor(offsetLeft))}px`,
+      overflow: "hidden",
+      [paddingProperty]: `${scrollbarWidth}px`,
+      position: "fixed",
+      right: "0",
+      top: `${-(win.scrollY - Math.floor(offsetTop))}px`,
+    })
+
+    return () => {
+      restoreStyle()
+      win.scrollTo({ behavior: "instant", left: win.scrollX, top: win.scrollY })
+    }
   }
 
-  const { localName } = el
-  const focusableTags = ["input", "select", "textarea", "button"]
+  const unsubscribes = [
+    setStyle(doc.documentElement, {
+      "--scrollbar-width": `${scrollbarWidth}px`,
+    }),
+    isIos()
+      ? setStyleForIOS()
+      : setStyle(doc.body, {
+          overflow: "hidden",
+          [paddingProperty]: `${scrollbarWidth}px`,
+        }),
+  ]
 
-  if (focusableTags.includes(localName)) return true
+  return () => {
+    unsubscribes.forEach((unsubscribe) => unsubscribe())
 
-  const others = {
-    a: () => el.hasAttribute("href"),
-    audio: () => el.hasAttribute("controls"),
-    video: () => el.hasAttribute("controls"),
+    doc.body.removeAttribute(LOCK_DATA_ATTR)
   }
-
-  if (localName in others) return others[localName as keyof typeof others]()
-
-  if (isContentEditable(el)) return true
-
-  return hasTabIndex(el)
-}
-
-export function hasNegativeTabIndex(el: HTMLElement): boolean {
-  return hasTabIndex(el) && el.tabIndex === -1
-}
-
-export function isTabbable(el?: HTMLElement | null): boolean {
-  return el
-    ? isHTMLElement(el) && isFocusable(el) && !hasNegativeTabIndex(el)
-    : false
-}
-
-export function isTouchDevice(): boolean {
-  return "ontouchstart" in window
-}
-
-export function getOwnerWindow(
-  node?: Element | null,
-): typeof globalThis & Window {
-  return getOwnerDocument(node).defaultView ?? window
-}
-
-export function getOwnerDocument(el?: Element | null): Document {
-  return isElement(el) ? el.ownerDocument : document
-}
-
-export function getActiveElement(el?: HTMLElement): HTMLElement | null {
-  return getOwnerDocument(el).activeElement as HTMLElement | null
-}
-
-export function isActiveElement(el: HTMLElement): boolean {
-  return getActiveElement(el) === el
 }
