@@ -5,9 +5,9 @@ export * from "./color"
 export * from "./dom"
 export * from "./event"
 export * from "./function"
+export * from "./i18n"
 export type * from "./index.types"
 export * from "./module"
 export * from "./number"
 export * from "./object"
-export * from "./react"
 export * from "./string"
