@@ -7,84 +7,66 @@ import {
   isUndefined,
 } from "./assertion"
 
-function omitObjectHelper<Y extends Dict, M extends keyof Y>(
+export function omitObject<Y extends Dict, M extends keyof Y>(
   obj: Y,
-  path: M[] | readonly M[],
-): any {
-  if (!path.length) return obj
+  keys: M[] | readonly M[],
+): Omit<Y, M> {
+  if (!keys.length) return obj
 
-  const [primaryKey, ...restKeys] = path
+  const result: Dict = {}
 
-  if (restKeys.length === 0 && primaryKey && primaryKey in obj) {
-    const { [primaryKey]: _, ...rest } = obj
+  Object.keys(obj).forEach((key) => {
+    if (keys.includes(key as M)) return
 
-    return rest
-  }
+    result[key] = obj[key]
+  })
 
-  if (primaryKey && obj[primaryKey] && isObject(obj[primaryKey])) {
-    return {
-      ...obj,
-      [primaryKey]: omitObjectHelper(obj[primaryKey], restKeys),
-    }
-  }
-
-  return obj
+  return result as Omit<Y, M>
 }
 
-export function omitObject<
-  Y extends Dict,
-  M extends keyof Y = keyof Y,
-  D = any,
->(obj: Y, keys: M[] | readonly M[]) {
-  return (keys as M[]).reduce((prev, key) => {
-    const path = (isString(key) ? key.split(".") : []) as M[]
+export function pickObject<Y extends Dict, M extends keyof Y>(
+  obj: Y,
+  keys: M[] | readonly M[],
+): { [D in M]: Y[D] } {
+  if (!keys.length) return {} as { [D in M]: Y[D] }
 
-    return omitObjectHelper(prev, path)
-  }, obj) as unknown as D extends unknown ? Omit<Y, M> : D
-}
+  const result = {} as { [D in M]: Y[D] }
 
-export function pickObject<
-  Y extends Dict,
-  M extends keyof Y = keyof Y,
-  D = any,
->(obj: Y, keys: M[] | readonly M[], fallbackValue = "__fallback") {
-  return (keys as M[]).reduce((prev, key) => {
-    const path = isString(key) ? key.split(".") : []
+  keys.forEach((key) => {
+    if (key in obj) result[key] = obj[key]
+  })
 
-    if (!path.length) return prev
-
-    const value = getMemoizedObject(obj, key as string, fallbackValue)
-
-    if (value === fallbackValue) return prev
-
-    prev = merge(
-      prev,
-      path.reduceRight(
-        (prev, key) => ({ [key]: key === path.at(-1) ? value : prev }),
-        {},
-      ),
-    )
-
-    return prev
-  }, {}) as D extends unknown ? { [H in M]: Y[H] } : D
+  return result
 }
 
 export function splitObject<Y extends Dict, M extends keyof Y>(
   obj: Y,
-  keys: M[] | readonly M[],
+  funcOrKeys?: M[] | readonly M[],
+): [{ [D in M]: Y[D] }, Omit<Y, M>]
+
+export function splitObject<Y extends Dict, M extends Dict>(
+  obj: Dict,
+  funcOrKeys?: (key: string) => boolean,
+): [Y, M]
+
+export function splitObject<Y extends Dict, M extends keyof Y>(
+  obj: Y,
+  funcOrKeys?: ((key: string) => boolean) | M[] | readonly M[],
 ) {
   const picked: Dict = {}
   const omitted: Dict = {}
 
   for (const [key, value] of Object.entries(obj)) {
-    if (keys.includes(key as Y[M])) {
+    if (
+      isFunction(funcOrKeys) ? funcOrKeys(key) : funcOrKeys?.includes(key as M)
+    ) {
       picked[key] = value
     } else {
       omitted[key] = value
     }
   }
 
-  return [picked, omitted] as [{ [P in M]: Y[P] }, Omit<Y, M>]
+  return [picked, omitted] as [{ [D in M]: Y[D] }, Omit<Y, M>]
 }
 
 export function filterObject<Y extends Dict, M extends Dict>(
@@ -106,26 +88,40 @@ export function filterUndefined<Y extends Dict>(obj: Y): Y {
   return filterObject(obj, (_, val) => val !== null && val !== undefined)
 }
 
+export interface mergeOptions {
+  debug?: boolean
+  mergeArray?: boolean
+  shouldProcess?: (value: any) => boolean
+}
+
 export function merge<Y extends Dict>(
   target: any,
   source: any,
-  mergeArray = false,
+  options: mergeOptions = { mergeArray: false, shouldProcess: () => true },
 ): Y {
   let result = Object.assign({}, target)
 
-  if (isObject(source)) {
+  if (isObject(source) && options.shouldProcess?.(source)) {
     if (isObject(target)) {
       for (const [sourceKey, sourceValue] of Object.entries(source)) {
         const targetValue: any = target[sourceKey]
 
-        if (mergeArray && isArray(sourceValue) && isArray(targetValue)) {
-          result[sourceKey] = targetValue.concat(...sourceValue)
-        } else if (
-          !isFunction(sourceValue) &&
-          isObject(sourceValue) &&
-          target.hasOwnProperty(sourceKey)
-        ) {
-          result[sourceKey] = merge(targetValue, sourceValue, mergeArray)
+        if (options.shouldProcess(sourceValue)) {
+          if (
+            options.mergeArray &&
+            isArray(sourceValue) &&
+            isArray(targetValue)
+          ) {
+            result[sourceKey] = targetValue.concat(...sourceValue)
+          } else if (
+            !isFunction(sourceValue) &&
+            isObject(sourceValue) &&
+            target.hasOwnProperty(sourceKey)
+          ) {
+            result[sourceKey] = merge(targetValue, sourceValue, options)
+          } else {
+            Object.assign(result, { [sourceKey]: sourceValue })
+          }
         } else {
           Object.assign(result, { [sourceKey]: sourceValue })
         }
@@ -140,32 +136,25 @@ export function merge<Y extends Dict>(
 
 export interface FlattenObjectOptions {
   maxDepth?: number
-  omitKeys?: string[]
   separator?: string
-  shouldProcess?: (obj: any) => boolean
+  shouldProcess?: (value: any) => boolean
 }
 
 export function flattenObject<Y extends Dict>(
   obj: any,
   {
     maxDepth = Infinity,
-    omitKeys = [],
     separator = ".",
-    shouldProcess,
+    shouldProcess = () => true,
   }: FlattenObjectOptions = {},
 ): Y {
-  if ((!isObject(obj) && !isArray(obj)) || !maxDepth) return obj
+  if (!isObject(obj) && !isArray(obj)) return obj
 
   return Object.entries(obj).reduce<any>((result, [key, value]) => {
-    if (
-      isObject(value) &&
-      !Object.keys(value).some((key) => omitKeys.includes(key)) &&
-      (!shouldProcess || shouldProcess(value))
-    ) {
+    if (isObject(value) && shouldProcess(value) && maxDepth > 1) {
       Object.entries(
         flattenObject(value, {
           maxDepth: maxDepth - 1,
-          omitKeys,
           separator,
           shouldProcess,
         }),
@@ -180,15 +169,7 @@ export function flattenObject<Y extends Dict>(
   }, {}) as Y
 }
 
-export function objectFromEntries<Y extends Dict>(entries: any[][]): Y {
-  return entries.reduce<any>((result, [key, value]) => {
-    result[key] = value
-
-    return result
-  }, {}) as Y
-}
-
-export function keysFormObject<Y extends object>(obj: Y): (keyof Y)[] {
+export function objectKeys<Y extends object>(obj: Y): (keyof Y)[] {
   return Object.keys(obj) as (keyof Y)[]
 }
 
@@ -213,18 +194,15 @@ export function getObject(
   obj: Dict | undefined,
   path: number | string,
   fallback?: any,
-  i?: number,
 ) {
   const keys = isString(path)
     ? path.split(/\[(.*?)\]|\./).filter(Boolean)
     : [path]
 
-  for (i = 0; i < keys.length; i += 1) {
+  for (const key of keys) {
     if (!obj) break
 
-    const key = keys[i]
-
-    obj = key ? obj[key] : undefined
+    obj = obj[key]
   }
 
   return obj === undefined ? fallback : obj
@@ -237,7 +215,6 @@ export function memoizeObject(func: typeof getObject) {
     obj: Dict,
     path: number | string,
     fallback?: any,
-    i?: number,
   ): Y {
     if (isUndefined(obj)) return func(obj, path, fallback)
 
@@ -247,7 +224,7 @@ export function memoizeObject(func: typeof getObject) {
 
     if (map.has(path)) return map.get(path)
 
-    const value = func(obj, path, fallback, i)
+    const value = func(obj, path, fallback)
 
     map.set(path, value)
 
@@ -259,20 +236,10 @@ export function memoizeObject(func: typeof getObject) {
 
 export const getMemoizedObject = memoizeObject(getObject)
 
-export function assignAfter(target: { [key: string]: any }, ...sources: any[]) {
-  const result: { [key: string]: unknown } = { ...target }
-
-  for (const nextSource of sources) {
-    if (nextSource == null) continue
-
-    for (const nextKey in nextSource) {
-      if (!Object.prototype.hasOwnProperty.call(nextSource, nextKey)) continue
-
-      if (nextKey in result) delete result[nextKey]
-
-      result[nextKey] = nextSource[nextKey]
-    }
+export function wrapWithKey<T>(obj: T, key?: string) {
+  if (key) {
+    return { [key]: obj }
+  } else {
+    return obj
   }
-
-  return result
 }
