@@ -1,4 +1,12 @@
-import type { StringLiteral } from "./index.types"
+import type { AnyString } from "./index.types"
+import { isString, isUndefined } from "./assertion"
+
+export function cx(...classNames: (string | undefined)[]) {
+  return classNames
+    .filter((className) => !isUndefined(className))
+    .map((className) => className.trim())
+    .join(" ")
+}
 
 export function escape(value: string, replaceValue = ""): string {
   return value.replace(/\s+/g, replaceValue)
@@ -59,7 +67,7 @@ export function antonym(value: string): string {
   }
 }
 
-const sizeMap: StringLiteral[] = [
+const sizeMap: AnyString[] = [
   "9xs",
   "8xs",
   "7xs",
@@ -71,7 +79,6 @@ const sizeMap: StringLiteral[] = [
   "xs",
   "sm",
   "md",
-  "normal",
   "lg",
   "xl",
   "2xl",
@@ -84,11 +91,15 @@ const sizeMap: StringLiteral[] = [
   "9xl",
 ]
 
-export function transformSize(
+export function isSize(value: any): boolean {
+  return isString(value) && sizeMap.includes(value)
+}
+
+export function transformSize<Y = string>(
   token: string | undefined,
   value: number,
-  omitTokens: null | string[] = ["normal"],
-): string | undefined {
+  omitTokens: null | string[] = null,
+): undefined | Y {
   if (!token) return undefined
 
   let resolvedSizeMap = sizeMap
@@ -99,28 +110,41 @@ export function transformSize(
   const index = resolvedSizeMap.indexOf(token)
   const size = resolvedSizeMap[index + value]
 
-  return size ?? token
+  return (size ?? token) as Y
 }
 
-export function toCamelCase(value: StringLiteral): string {
-  return value.toLowerCase().replace(/[_-](.)/g, (_, val) => val.toUpperCase())
+export function toCamelCase(value: AnyString): string {
+  return value
+    .replace(/[_-](.)/g, (_, val) => val.toUpperCase())
+    .replace(/^(.)/, (_, val) => val.toLowerCase())
 }
 
-export function toPascalCase(value: StringLiteral): string {
-  return toCamelCase(value).replace(/^(.)/, (_, val) => val.toUpperCase())
+export function toPascalCase(value: AnyString): string {
+  return value
+    .replace(/[_-](.)/g, (_, val) => val.toUpperCase())
+    .replace(/^(.)/, (_, val) => val.toUpperCase())
 }
 
-export function toKebabCase(value: StringLiteral): string {
+export function toKebabCase(value: AnyString): string {
   return value
     .replace(/([a-z0-9]|(?=[A-Z]))([A-Z])/g, "$1-$2")
     .toLowerCase()
     .replace(/^-/, "")
 }
 
-export function toTitleCase(value: StringLiteral): string {
+export function toTitleCase(value: AnyString): string {
   return value
     .replace(/([A-Z])/g, " $1")
     .replace(/[_-](.)/g, (_, val) => ` ${val.toUpperCase()}`)
     .replace(/^./, (str) => str.toUpperCase())
     .trim()
+}
+
+export function bem(block: string, element?: string, modifier?: string) {
+  let className = block
+
+  if (element) className += `__${element}`
+  if (modifier) className += `--${modifier}`
+
+  return className
 }
