@@ -2,7 +2,7 @@ import type { RestEndpointMethodTypes } from "@octokit/rest"
 import type { Project } from "find-packages"
 import * as p from "@clack/prompts"
 import { Octokit } from "@octokit/rest"
-import { isArray } from "@yamada-ui/react"
+import { isArray } from "@yamada-ui/utils"
 import c from "chalk"
 import { findPackages } from "find-packages"
 import { existsSync } from "fs"
